requirements.txt
requests==2.31.0
tiktoken==0.5.2
python-dotenv==1.0.0
tqdm==4.66.1
markdownify==0.11.6
numpy==1.26.2
PyPDF2==3.0.1
python-docx==1.0.1 